Extract provider test cases
